backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
PyJWT